Add supporting classes and macros for error code and class decleration
Add macros and classes for automatically creating/declaring an error code, class, and associated category. These allow seamless use of a given error code enum or class as the error type for` result` without the usual boilerplate repitition.

Also reorganize the entire project/repo to fit the new files for this change.